The previous code would Marshal the header allocating (header) bytes.
Then call append with the payload allocating payload bytes and copying
(header) bytes. Then later EncryptRTP will append the auth tag,
potentially allocating (header + payload + 10) bytes and copying
(header + payload) bytes.

This code avoids the extra allocation/copy by allocating the correct
size upfront.

```
name                 old time/op    new time/op    delta
WriteRTP-8             9.73µs ± 2%    9.33µs ± 2%   -4.11%

name                 old alloc/op   new alloc/op   delta
WriteRTP-8             3.17kB ± 0%    1.10kB ± 0%  -65.15%

name                 old allocs/op  new allocs/op  delta
WriteRTP-8               13.0 ± 0%      10.0 ± 0%  -23.08%
```

The benchmark prints four `io: read/write on closed pipe` log lines but I wasn't able to figure it how to fix them. It's pretty annoying but I'll fix it later.

Depends on https://github.com/pions/rtp/pull/3