@kixelated 2019-02-25
This change destructively overwrites the buffer, which is not supposed to be done by `io.Writer`s. Alongside the need to make the buffer reuse thread-safe, there's really no performance improvement any longer. Just one less allocation the size of the header so really not worth the effort to do anything but CLOSE.